# modules/schema/openapi.json
{
  "components": {
...
    "schemas": {
      "Pick_PhabricatorAccountEntity.id-or-email-or-url-or-apiKey_": {
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "url",
          "apiKey"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "PhabricatorAccountState": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Pick_PhabricatorAccountEntity.id-or-email-or-url-or-apiKey_"
          },
          {
            "properties": {
              "enabledQueries": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              "icon": {
                "type": "string"
              }
            },
            "required": [
              "enabledQueries",
              "icon"
            ],
            "type": "object"
          }
        ]
      },
      "Pick_PhabricatorAccountEntity.url-or-apiKey_": {
        "properties": {
          "url": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "apiKey"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "PhabricatorAccountParams": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Pick_PhabricatorAccountEntity.url-or-apiKey_"
          },
          {
            "properties": {
              "queries": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            },
            "required": [
              "queries"
            ],
            "type": "object"
          }
        ]
      },
      "Partial_PhabricatorAccountParams_": {
        "properties": {
          "url": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          },
          "queries": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object",
        "description": "Make all properties in T optional"
      },
...
      "Pick_JiraAccountEntity.Exclude_keyofJiraAccountEntity.userId__": {
        "properties": {
          "id": {
            "type": "string"
          },
          "apiToken": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "serverName": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "apiToken",
          "email",
          "serverName",
          "url",
          "userName"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_JiraAccountEntity.userId_": {
        "$ref": "#/components/schemas/Pick_JiraAccountEntity.Exclude_keyofJiraAccountEntity.userId__"
      },
      "Pick_JiraSearchEntity.Exclude_keyofJiraSearchEntity.accountId__": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "dueOffset": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "query",
          "dueOffset"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_JiraSearchEntity.accountId_": {
        "$ref": "#/components/schemas/Pick_JiraSearchEntity.Exclude_keyofJiraSearchEntity.accountId__"
      },
      "JiraSearchState": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Omit_JiraSearchEntity.accountId_"
          },
          {
            "properties": {
              "url": {
                "type": "string"
              }
            },
            "required": [
              "url"
            ],
            "type": "object"
          }
        ]
      },
      "JiraAccountState": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Omit_JiraAccountEntity.userId_"
          },
          {
            "properties": {
              "searches": {
                "items": {
                  "$ref": "#/components/schemas/JiraSearchState"
                },
                "type": "array"
              }
            },
            "required": [
              "searches"
            ],
            "type": "object"
          }
        ]
      },
      "Pick_JiraAccountEntity.Exclude_keyofJiraAccountEntity.id-or-userId-or-serverName-or-userName__": {
        "properties": {
          "apiToken": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "apiToken",
          "email",
          "url"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_JiraAccountEntity.id-or-userId-or-serverName-or-userName_": {
        "$ref": "#/components/schemas/Pick_JiraAccountEntity.Exclude_keyofJiraAccountEntity.id-or-userId-or-serverName-or-userName__"
      },
      "JiraAccountParams": {
        "$ref": "#/components/schemas/Omit_JiraAccountEntity.id-or-userId-or-serverName-or-userName_"
      },
...
      "Pick_BugzillaAccountEntity.id-or-name-or-icon-or-url_": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "icon",
          "url"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "SearchType": {
        "enum": [
          "quick",
          "advanced"
        ],
        "type": "string",
        "x-enum-varnames": [
          "Quicksearch",
          "Advanced"
        ]
      },
      "Pick_BugzillaSearchEntity.Exclude_keyofBugzillaSearchEntity.accountId__": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/SearchType"
          },
          "dueOffset": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "query",
          "type",
          "dueOffset"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_BugzillaSearchEntity.accountId_": {
        "$ref": "#/components/schemas/Pick_BugzillaSearchEntity.Exclude_keyofBugzillaSearchEntity.accountId__"
      },
      "BugzillaSearchState": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Omit_BugzillaSearchEntity.accountId_"
          },
          {
            "properties": {
              "url": {
                "type": "string"
              }
            },
            "required": [
              "url"
            ],
            "type": "object"
          }
        ]
      },
      "BugzillaAccountState": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Pick_BugzillaAccountEntity.id-or-name-or-icon-or-url_"
          },
          {
            "properties": {
              "searches": {
                "items": {
                  "$ref": "#/components/schemas/BugzillaSearchState"
                },
                "type": "array"
              }
            },
            "required": [
              "searches"
            ],
            "type": "object"
          }
        ]
      },
      "Pick_BugzillaAccountEntity.Exclude_keyofBugzillaAccountEntity.id-or-userId-or-icon__": {
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "url",
          "username",
          "password"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_BugzillaAccountEntity.id-or-userId-or-icon_": {
        "$ref": "#/components/schemas/Pick_BugzillaAccountEntity.Exclude_keyofBugzillaAccountEntity.id-or-userId-or-icon__"
      },
      "BugzillaAccountParams": {
        "$ref": "#/components/schemas/Omit_BugzillaAccountEntity.id-or-userId-or-icon_"
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": ""
  },
  "openapi": "3.0.0",
  "paths": {
    "/service/phabricator/account": {
      "get": {
        "operationId": "ListAccounts",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PhabricatorAccountState"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": []
      },
      "put": {
        "operationId": "CreateAccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhabricatorAccountState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "params": {
                    "$ref": "#/components/schemas/PhabricatorAccountParams"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "EditAccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhabricatorAccountState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "params": {
                    "$ref": "#/components/schemas/Partial_PhabricatorAccountParams_"
                  },
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "params",
                  "id"
                ],
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DeleteAccount",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/service/jira/account": {
      "get": {
        "operationId": "ListAccounts",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JiraAccountState"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": []
      },
      "put": {
        "operationId": "CreateAccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraAccountState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "params": {
                    "$ref": "#/components/schemas/JiraAccountParams"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DeleteAccount",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/service/bugzilla/account": {
      "get": {
        "operationId": "ListAccounts",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BugzillaAccountState"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": []
      },
      "put": {
        "operationId": "CreateAccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BugzillaAccountState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "params": {
                    "$ref": "#/components/schemas/BugzillaAccountParams"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DeleteAccount",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}